requirements.txt
# ── FLUXION Python Dependencies ──────────────────────────────
# Install: pip install -r requirements.txt

# Core API Framework
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
pydantic>=2.10,<3.0

# HTTP Client (for Overpass/Geoapify calls)
httpx>=0.28,<1.0

# Database (PostgreSQL async + SQLAlchemy)
sqlalchemy[asyncio]>=2.0,<3.0
asyncpg>=0.30,<1.0

# Auth (JWT + password hashing)
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
bcrypt>=4.0,<5.0

# Environment variables
python-dotenv>=1.0,<2.0

# Starlette (middleware support, pulled by FastAPI but pinned)
starlette>=0.45,<1.0

# File Upload support (multipart)
python-multipart>=0.0.9,<1.0

# CSV / Excel import
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0

# Clustering for large-scale VRP (k-means)
scikit-learn>=1.4,<2.0

# Testing
pytest>=8.0,<9.0
pytest-asyncio>=0.24,<1.0

# Database Migrations
alembic>=1.14,<2.0
